move radio button and group to positronModalDialog components
