+++ src/vs/workbench/browser/positronComponents/positronModalDialog/components/radioButton.css
